Split Text into multiple files to allow for connect
