+++ frontend/react/src/components/fields/MailingAddress.js
@@ -1,4 +1,4 @@
-import { TextMultiline } from "./Text";
+import { TextMultiline } from "./TextOther";
 
 export { TextMultiline as MailingAddress };
 export default TextMultiline;
