+++ frontend/react/src/components/fields/Text.js
@@ -3,9 +3,9 @@ import PropTypes from "prop-types";
 import { TextField } from "@cmsgov/design-system-core";
 import ReactHtmlParser from "react-html-parser";
 import axios from "../../authenticatedAxios";
+import { connect } from "react-redux";
 
-const Text = ({ question, ...props }) => {
-  let a;
+const Text = ({ question, state, ...props }) => {
   const [printValue, setPrintValue] = useState(
     (question.answer && question.answer.entry) || ""
   );
@@ -24,35 +24,39 @@ const Text = ({ question, ...props }) => {
 
       // Even years get inputs, odd years get previous year data
       const shouldGetPriorYear = splitID[0] % 2;
-      let a = question.id;
-      let b;
+
+      // If question is Part 6, section 3c, question 9
       if (
-        !shouldGetPriorYear &&
+        shouldGetPriorYear &&
         splitID[1] === "03" &&
         splitID[2] === "c" &&
         splitID[3] === "06" &&
         splitID[4] === "09"
       ) {
-        setPrevYearDisabled(true);
         // Set year to last year
         const lastYear = parseInt(splitID[0]) - 1;
         splitID[0] = lastYear;
         const fieldsetId = splitID.join("-");
 
-        const state = "AK";
-
-        await axios
-          .get(`/api/v1/sections/${lastYear}/${state.toUpperCase()}/3`)
-          .then((data) => {
-            console.log("zzzData", data);
-            let a;
-            // getDataFromLastYear(fieldsetId);
-            if (data) {
-              let value = getValueFromLastYear(data.data, fieldsetId) || "";
-              let p;
-              setPrevYearValue(value);
-            }
-          });
+        if (state) {
+          try {
+            await axios
+              .get(`/api/v1/sections/${lastYear}/${state.toUpperCase()}/3`)
+              .then((data) => {
+                // getDataFromLastYear(fieldsetId);
+                if (data) {
+                  let value =
+                    getValueFromLastYear(data.data, fieldsetId) ||
+                    "No Value from Last Year";
+                  setPrevYearDisabled(true);
+                  setPrevYearValue(value);
+                }
+              });
+          } catch (e) {
+            setPrevYearDisabled(true);
+            setPrevYearValue(null);
+          }
+        }
       }
     };
 
@@ -60,24 +64,22 @@ const Text = ({ question, ...props }) => {
       let lastYearAnswer;
       // Get questions from last years JSON
       const questions = data.contents.section.subsections[2].parts[5].questions;
-      let a;
+
+      // Derive matching question
       let matchingQuestion = questions.filter(
         (question) => fieldsetId === question?.id
       );
-      let b;
-      if (matchingQuestion[0]) {
-        // Since these always go in order we get the subquestion ID, convert to lowercase letter, get the char code (a = 97)
-        // and subtract 97 to get the question index number
 
-        let c;
+      // Always use first item in array
+      if (matchingQuestion[0]) {
         lastYearAnswer = matchingQuestion[0].answer.entry;
       }
-      let d;
-      return lastYearAnswer ?? 0;
+
+      return lastYearAnswer ?? "";
     };
 
     getPrevYearValue().then();
-  }, []);
+  }, [state]);
 
   return (
     <>
@@ -97,20 +99,11 @@ const Text = ({ question, ...props }) => {
 };
 Text.propTypes = {
   question: PropTypes.object.isRequired,
+  state: PropTypes.string.isRequired,
 };
 
-const TextMedium = ({ question, ...props }) => (
-  <Text question={question} multiline rows={3} {...props} />
-);
-TextMedium.propTypes = {
-  question: PropTypes.object.isRequired,
-};
+const mapState = (state) => ({
+  state: state.formData[0].contents?.section?.state,
+});
 
-const TextMultiline = ({ question, ...props }) => (
-  <Text question={question} multiline rows={6} {...props} />
-);
-TextMultiline.propTypes = {
-  question: PropTypes.object.isRequired,
-};
-
-export { Text, TextMedium, TextMultiline, Text as TextSmall };
+export default connect(mapState)(Text);
